+++ src/app/dashboard/page.tsx
@@ -1,4 +1,4 @@
-import { AppSidebar } from "@/components/app-sidebar";
+import { AppSidebar } from "@/components/sidebar/app-sidebar";
 import Header from "@/components/header/Header";
 import {
   Breadcrumb,
